* release/1.0.0: (43 commits)
  Put back include path for now
  Remove hard-coded path from tcp script
  Re-introduce debug messages
  Stop using legacy archiver in FDB5Sink; remove FdbDirectSink
  MUL-14: ensure MIO is contructed
  Change path for JSON.h
  multio-server: Remove includes unnecessary with CMake 3.0
  Remove multio-grib
  version bump
  MUL-14 remove all interface
  Sink output from IFS to files
  multio-server: Update IFS interface to send flushes and fields
  multio-server: re-organise multio-hammer
  multio-server: make Peer support different dynamic types
  multio-server: Aggregate flush messages
  multio-server: Distinguish between notification and flush message
  Add sink action to sink each field to a separate file
  multio-server: Split PlainBlob between header and cc
  multio-hammer: Read list of parameters from configuration
  multio-server: Notification when step_complete message is received
  ...

+++ src/multio/server/LocalIndices.cc
@@ -8,7 +8,7 @@
 namespace multio {
 namespace server {
 
-    LocalIndices::LocalIndices(std::vector<size_t>&& idx) : indices_(std::move(idx)) {}
+LocalIndices::LocalIndices(std::vector<int32_t>&& idx) : indices_(std::move(idx)) {}
 
 void LocalIndices::to_global(const std::vector<double>& local, std::vector<double>& global) const {
     ASSERT(local.size() == indices_.size());
@@ -21,7 +21,7 @@ void LocalIndices::to_global(const std::vector<double>& local, std::vector<doubl
 
 void LocalIndices::to_local(const std::vector<double>& global, std::vector<double>& local) const {
     local.resize(0);
-    std::for_each(begin(indices_), end(indices_), [&](size_t id) { local.push_back(global[id]); });
+    std::for_each(begin(indices_), end(indices_), [&](int32_t id) { local.push_back(global[id]); });
 }
