+++ src/multio/server/LocalIndices.h
@@ -3,6 +3,7 @@
 #define multio_server_LocalIndices_H
 
 #include <cstddef>
+#include <cstdint>
 #include <vector>
 
 namespace multio {
@@ -10,13 +11,13 @@ namespace server {
 
 class LocalIndices {
 public:
-    LocalIndices(std::vector<size_t>&& idx);
+    LocalIndices(std::vector<int32_t>&& idx);
 
     void to_global(const std::vector<double>& local, std::vector<double>& global) const;
     void to_local(const std::vector<double>& global, std::vector<double>& local) const;
 
 private:
-    std::vector<size_t> indices_;  // Unstructured
+    std::vector<int32_t> indices_;  // Grid-point
 };
 
 }  // namespace server
